fix bug with transformed bounds functions and orientation: YAY all the tests in cairo/test now work!!!
The long version: computing a transformed bounding function involves
taking a vector v, computing a set of vectors spanning the hyperplane
orthogonal to v, transforming the hyperplane, then finding a vector
orthogonal to the transformed hyperplane.  The bug was basically that
I failed to realize the last step can have two essentially different
solutions, one pointing in each direction.  But only one of them is
actually correct, given the vector we started with.  The solution is
to remember the original vector as the "orientation" of the
hyperplane; then when generating the vector orthogonal to the
transformed hyperplane we check whether it is to the same side of the
hyperplane as the (transformed) orientation vector, and negate it if
not.