New functions envelope[VP]May
envelopeV and envelopeP return the zero vector and origin,
respectively, when called on an empty envelope.  However, sometimes
it's useful to actually know whether the envelope was empty or not
(the zero vector and the origin are legitimate outputs from non-empty
envelopes).  The new functions have their return type wrapped in Maybe
for this purpose.