Depend on vector-space-0.8.4 and remove some redundant constraints
vector-space-0.8.4 adds AdditiveGroup (Scalar v) as a superclass of
InnerSpace v.  So we can now remove AdditiveGroup (Scalar v)
constraints in a few places where we already have an InnerSpace v
constraint.

Unfortunately, there are a lot of places where we actually have
OrderedField (Scalar v).  I rather suspect that many of these are
overkill, and that perhaps we ought to do away with it and just list
out the constraints...