Generalize Transformable instances for pairs and triples
Previously, the components of the tuples were required to have the
same type; but everything still works as long as they all share the
same vector space.  This is actually useful in practice: say, if we
wanted to pair a diagram with a path and then apply the same
transformation to both.