bug fix: compute diameter correctly: add scalars instead of subtracting
Fixes #35. This one was sneaky.  The bug was introduced in
https://github.com/diagrams/diagrams-core/commit/0681b24b3d5c7294f2a960fb5d76eee5e6f378ec,
which looks innocuous on the surface -- but there's a big difference
between subtracting two *vectors* (what the original code did) and
subtracting two *scalars* (which is what the code does now).  What we
actually want to do is *add* the two scalars.