D.C.Compile: bug fix: act on styles when pushing transforms past them down the tree
Transformations implicitly have an action on everything below them in
a tree.  So when pushing a transformation down past a style node in a
tree we have to also use the transform to act on the style.  Right now
the only transformable attribute we have is clipping paths, so this
only showed up as a bug in certain cases when clipping was involved.