Move Context back to Core/Types.hs
This is needed because Context depends on Environment which in turn
depends on RTree. I'm not really sure how to fix this cleanly,
especially because Contextual was originally factored out to make
contextual envelopes possible.

Perhaps RTree+Summary needs to be factored out as well? I'll leave this
undecided for now.