Degeneralize Annotation to cut down on the number of type parameters.
It's not obvious to me if Annotation was intended to be polymorphic. If
this ends up breaking anything, it can be reverted but then an extra
annotation needs to be added to contexts. This will probably not be an
issue either way if Contextual is fixed as per the previous commit's
comment.